centralisedgraphcolourer
